PIC-2655 Ensure Offences API is only called for Magistrate Court hearings
